refactor(ui): split up base hooks into logical files
Move the base hooks into files related to their purpose.

Fixes: canonical-web-and-design/app-tribe#575.
