+++ ui/src/app/machines/components/MachineHeaderForms/ActionFormWrapper/DeployForm/DeployForm.test.tsx
@@ -6,7 +6,7 @@ import configureStore from "redux-mock-store";
 
 import DeployForm from "./DeployForm";
 
-import * as hooks from "app/base/hooks";
+import * as hooks from "app/base/hooks/analytics";
 import { PodType } from "app/store/pod/constants";
 import type { RootState } from "app/store/root/types";
 import { NodeActions } from "app/store/types/node";
